# Backend/api/documentation/openapi.yaml
openapi: 3.0.3
info:
  title: 'Kebab Api'
  description: ''
  version: 1.0.0
servers:
  -
    url: api/documentation
paths:
...
  '/api/kebabs/{kebabId}/name':
    put:
      summary: ''
      operationId: putApiKebabsKebabIdName
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: ''
                  example: molestiae
                  nullable: false
              required:
                - name
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: minima
        required: true
        schema:
          type: string
  '/api/kebabs/{kebabId}/address':
    put:
      summary: ''
      operationId: putApiKebabsKebabIdAddress
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: hasoytxcdkunpxsulsofhjcbd
                  nullable: false
              required:
                - address
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: aut
        required: true
        schema:
          type: string
  '/api/kebabs/{kebabId}/coordinates':
    put:
      summary: ''
      operationId: putApiKebabsKebabIdCoordinates
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coordinates:
                  type: object
                  description: ''
                  example: []
                  nullable: false
                  properties: {  }
              required:
                - coordinates
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: autem
        required: true
        schema:
          type: string
  '/api/kebabs/{kebabId}/sauce':
    post:
      summary: ''
      operationId: postApiKebabsKebabIdSauce
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sauce:
                  type: string
                  description: ''
                  example: nulla
                  nullable: false
              required:
                - sauce
      security: []
    delete:
      summary: ''
      operationId: deleteApiKebabsKebabIdSauce
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sauce:
                  type: string
                  description: ''
                  example: est
                  nullable: false
              required:
                - sauce
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: accusamus
        required: true
        schema:
          type: string
  '/api/kebabs/{kebabId}/meat':
    post:
      summary: ''
      operationId: postApiKebabsKebabIdMeat
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                meat:
                  type: string
                  description: ''
                  example: voluptates
                  nullable: false
              required:
                - meat
      security: []
    delete:
      summary: ''
      operationId: deleteApiKebabsKebabIdMeat
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                meat:
                  type: string
                  description: ''
                  example: suscipit
                  nullable: false
              required:
                - meat
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: et
        required: true
        schema:
          type: string
  '/api/kebabs/{kebabId}/status':
    put:
      summary: ''
      operationId: putApiKebabsKebabIdStatus
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: exists
                  nullable: false
                  enum:
                    - exists
                    - closed
                    - planned
              required:
                - status
      security: []
    parameters:
      -
        in: path
        name: kebabId
        description: ''
        example: doloribus
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []